Moved unit tests that are integration tests
