+++ prime-router/src/test/kotlin/FakeReportTests.kt
@@ -1,9 +1,7 @@
 package gov.cdc.prime.router
 
 import assertk.assertThat
-import assertk.assertions.contains
 import assertk.assertions.isEqualTo
-import assertk.assertions.isNotNull
 import assertk.assertions.isTrue
 import gov.cdc.prime.router.metadata.LookupTable
 import gov.cdc.prime.router.unittest.UnitTestUtils
@@ -196,43 +194,5 @@ class FakeReportTests {
         val actual = FakeReport(metadata).buildMappedColumn(concatField, rowContext)
         val expected = "Any lab USA, Any facility USA"
         assertThat(actual).isEqualTo(expected)
-    }
-
-    @Test
-    fun `test row context getting expected zip code results`() {
-        // arrange
-        val metadata = UnitTestUtils.testMetadata
-        assertThat(metadata).isNotNull()
-        val zipCodeTable = metadata.findLookupTable("zip-code-data")
-        assertThat(zipCodeTable).isNotNull()
-        val state = "VT"
-        val county = "Rutland"
-        val matchingCityRows = zipCodeTable!!.filter(
-            "city",
-            mapOf(
-                "state_abbr" to state,
-                "county" to county
-            )
-        )
-        val matchingZipRows = zipCodeTable.filter(
-            "zipcode",
-            mapOf(
-                "state_abbr" to state,
-                "county" to county
-            )
-        )
-        // act
-        val context = FakeReport.RowContext(
-            metadata::findLookupTable,
-            state,
-            null,
-            county
-        )
-        println(matchingCityRows.joinToString())
-        println(matchingZipRows.joinToString())
-        println("${context.city} - ${context.zipCode}")
-        // assert
-        assertThat(matchingCityRows).contains(context.city)
-        assertThat(matchingZipRows).contains(context.zipCode)
     }
 }
