+++ prime-router/src/testIntegration/kotlin/FileNameTemplateIntegrationTests.kt
@@ -1,5 +1,3 @@
-package gov.cdc.prime.router
-
 import assertk.Assert
 import assertk.assertThat
 import assertk.assertions.isEqualTo
@@ -13,7 +11,16 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import com.fasterxml.jackson.module.kotlin.readValue
-import gov.cdc.prime.router.unittest.UnitTestUtils
+import gov.cdc.prime.router.CreatedDate
+import gov.cdc.prime.router.FileNameTemplate
+import gov.cdc.prime.router.Hl7Configuration
+import gov.cdc.prime.router.Literal
+import gov.cdc.prime.router.Metadata
+import gov.cdc.prime.router.ProcessingModeCode
+import gov.cdc.prime.router.ReceivingOrganization
+import gov.cdc.prime.router.RegexReplace
+import gov.cdc.prime.router.SchemaBaseName
+import gov.cdc.prime.router.TranslatorConfiguration
 import io.mockk.every
 import io.mockk.mockkClass
 import java.time.OffsetDateTime
@@ -22,7 +29,7 @@ import java.util.UUID
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 
-class FileNameTemplateTests {
+class FileNameTemplateIntegrationTests {
     private val literal = """
         ---
         elements:
@@ -39,7 +46,7 @@ class FileNameTemplateTests {
         every { it.receivingOrganization }.returns("yoyodyne")
     }
     private val mapper = ObjectMapper(YAMLFactory()).registerModule(KotlinModule())
-    private val metadata = UnitTestUtils.testMetadata
+    private val metadata = Metadata.getInstance()
     private val dateFormat = "yyyyMMdd"
     private val formatter = DateTimeFormatter.ofPattern(dateFormat)
     private val reportId = UUID.randomUUID()
